+++ InstaBotPrototype/InstaBotLibrary/Models/ConfigurationModel.cs
@@ -1,0 +1,11 @@
+﻿namespace InstaBotPrototype.Models
+{
+    public class ConfigurationModel
+    {
+        public string InstaUsername { get; set; }
+        public string InstaPassword { get; set; }
+        public string TelegramUsername { get; set; }
+        public string Tags { get; set; }
+        public string Topics { get; set; }
+    }
+}
