+++ InstaBotPrototype/InstaBotLibrary/Services/Telegram/ITelegramService.cs
@@ -1,0 +1,8 @@
+﻿namespace InstaBotPrototype.Services.Telegram
+{
+    public interface ITelegramService
+    {
+        int Connect(string username);
+        string SendMessage(string message);
+    }
+}
