+++ InstaBotPrototype/InstaBotPrototype.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26430.13
+VisualStudioVersion = 15.0.27004.2009
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InstaBotPrototype", "InstaBotPrototype\InstaBotPrototype.csproj", "{11FFC5F1-984D-440C-AA2D-9306FEE59D7B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InstaBotPrototype", "InstaBotPrototype\InstaBotPrototype.csproj", "{11FFC5F1-984D-440C-AA2D-9306FEE59D7B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InstaBotLibrary", "InstaBotLibrary\InstaBotLibrary.csproj", "{33417DF4-E028-47B3-887B-12F59BD2ED44}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,8 +17,15 @@ Global
 		{11FFC5F1-984D-440C-AA2D-9306FEE59D7B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{11FFC5F1-984D-440C-AA2D-9306FEE59D7B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{11FFC5F1-984D-440C-AA2D-9306FEE59D7B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{33417DF4-E028-47B3-887B-12F59BD2ED44}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{33417DF4-E028-47B3-887B-12F59BD2ED44}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{33417DF4-E028-47B3-887B-12F59BD2ED44}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{33417DF4-E028-47B3-887B-12F59BD2ED44}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {DE2BEE13-65D4-4D6A-85F1-5B8CCA3A8B3E}
 	EndGlobalSection
 EndGlobal
