--- InstaBotPrototype/InstaBotPrototype/Services/Telegram/ITelegramService.cs
@@ -1,8 +1,0 @@
-﻿namespace InstaBotPrototype.Services.Telegram
-{
-    public interface ITelegramService
-    {
-        int Connect(string username);
-        string SendMessage(string message);
-    }
-}
